Switch to dynamic plugin: @backstage/plugin-catalog-backend-module-github
