+++ dynamic-plugins/plugin-catalog-backend-module-github-wrapped/.eslintignore
@@ -1,0 +1,1 @@
+dist-dynamic
