+++ dynamic-plugins/plugin-catalog-backend-module-github-wrapped/.eslintrc.js
@@ -1,0 +1,1 @@
+module.exports = require('@backstage/cli/config/eslint-factory')(__dirname);
