+++ dynamic-plugins/plugin-catalog-backend-module-github-wrapped/dist-dynamic/package.json
@@ -1,0 +1,69 @@
+{
+  "name": "plugin-catalog-backend-module-github-wrapped-dynamic",
+  "version": "0.4.3-dynamic.0",
+  "main": "dist/index.cjs.js",
+  "types": "dist/index.d.ts",
+  "license": "Apache-2.0",
+  "publishConfig": {
+    "access": "public",
+    "main": "dist/index.cjs.js",
+    "types": "dist/index.d.ts"
+  },
+  "backstage": {
+    "role": "backend-plugin-module"
+  },
+  "exports": {
+    ".": {
+      "require": "./dist/index.cjs.js",
+      "default": "./dist/index.cjs.js"
+    },
+    "./package.json": "./package.json"
+  },
+  "scripts": {
+    "build": "backstage-cli package build",
+    "lint": "backstage-cli package lint",
+    "test": "backstage-cli package test --coverage",
+    "clean": "backstage-cli package clean",
+    "clean-dynamic-sources": "yarn clean && rm -Rf node_modules",
+    "export-dynamic": "janus-cli package export-dynamic-plugin --embed-package @backstage/plugin-catalog-backend-module-github"
+  },
+  "dependencies": {
+    "@octokit/graphql": "^5.0.0",
+    "@octokit/rest": "^19.0.3",
+    "git-url-parse": "^13.0.0",
+    "lodash": "^4.17.21",
+    "minimatch": "^5.1.2",
+    "node-fetch": "^2.6.7",
+    "uuid": "^8.0.0",
+    "winston": "^3.2.1"
+  },
+  "devDependencies": {},
+  "files": [
+    "dist"
+  ],
+  "bundleDependencies": true,
+  "peerDependencies": {
+    "@backstage/backend-plugin-manager": "npm:@janus-idp/backend-plugin-manager@0.0.5-janus.0",
+    "@backstage/backend-common": "^0.19.7",
+    "@backstage/backend-plugin-api": "^0.6.5",
+    "@backstage/backend-tasks": "^0.5.10",
+    "@backstage/catalog-client": "^1.4.4",
+    "@backstage/catalog-model": "^1.4.2",
+    "@backstage/config": "^1.1.0",
+    "@backstage/errors": "^1.2.2",
+    "@backstage/integration": "^1.7.0",
+    "@backstage/plugin-catalog-backend": "^1.13.3",
+    "@backstage/plugin-catalog-common": "^1.0.16",
+    "@backstage/plugin-catalog-node": "^1.4.6",
+    "@backstage/plugin-events-node": "^0.2.14",
+    "@backstage/types": "^1.1.1"
+  },
+  "overrides": {
+    "@aws-sdk/util-utf8-browser": {
+      "@smithy/util-utf8": "^2.0.0"
+    }
+  },
+  "resolutions": {
+    "@aws-sdk/util-utf8-browser": "npm:@smithy/util-utf8@^2.0.0"
+  }
+}
