+++ dynamic-plugins/plugin-catalog-backend-module-github-wrapped/src/index.ts
@@ -1,0 +1,1 @@
+export * from './dynamic/index';
