+++ packages/backend/package.json
@@ -26,7 +26,6 @@
     "@backstage/plugin-auth-node": "0.3.2",
     "@backstage/plugin-azure-devops-backend": "0.4.2",
     "@backstage/plugin-catalog-backend": "1.13.3",
-    "@backstage/plugin-catalog-backend-module-github": "0.4.3",
     "@backstage/plugin-catalog-backend-module-openapi": "0.1.22",
     "@backstage/plugin-jenkins-backend": "0.2.8",
     "@backstage/plugin-kubernetes-backend": "0.12.2",
